Added algorithm to determine point is in a polygon
Added an algorithm to determine if a point is within a polygon, currently
it works well with all of the test cases I can think of. It treats any point
directly on the lines between vertices or vertex themselves as points within
the polygon.

The algorithm is a much simpler and elegant solution to the typic ray casting
and winding number solutions (and has better performance!). The algorithm
PNPOLY was created by WRF.

SEE: http://www.ecse.rpi.edu/Homepages/wrf/Research/Short_Notes/pnpoly.html