MODIFIED: Fixed convex check
Convex check now properly identifies whether the newly added point is connected in such a way that it would no longer be a convex hull. If the point's connection is a concave connection it will discard the connection point and attempt to connect to the next point and check the convex hull's integrety.